Fixed up look_in_room(), as the code was pretty messy.
Fixed sector type display to use the room that was passed in to look_in_room() instead of the room of the character that is looking -- this fixes sector type display when using Clairvoyance, etc. This was also broken previously, but probably not noticed since immortals virtually always use "at <where> look".

Added PLR_SECTOR_BRIEF for sector type, so that it has its own brief mode. There were some that asked for this, and before this I had just hard-wired it to not show when both normal and super brief modes were on, which wasn't very elegant.

Cleaned up some of structs.h -- it's in bad shape formatting/readability wise, so this will take a long time. There are a few sections of commented out code, as I am working on a few things for the future.

Added help entry info for sector brief.