Add uber Xy
Fix long descs in Sword of Truth

Removed trailing whitespace from "The Wight " in spell_animate_dead().

Changes Evasion to only function on physical and skill damage.

Added keyword 'mcorpse' for NPC (mcorpse just seems to sound better to a mudder than ncorpse) and 'pcorpse' for PC corpses.

Fixed up look_in_room(), as the code was pretty messy.

Fixed sector type display to use the room that was passed in to look_in_room() instead of the room of the character that is looking -- this fixes sector type display when using Clairvoyance, etc. This was also broken previously, but probably not noticed since immortals virtually always use "at <where> look".

Added PLR_SECTOR_BRIEF for sector type, so that it has its own brief mode. There were some that asked for this, and before this I had just hard-wired it to not show when both normal and super brief modes were on, which wasn't very elegant.

Cleaned up some of structs.h -- it's in bad shape formatting/readability wise, so this will take a long time. There are a few sections of commented out code, as I am working on a few things for the future.

Added help entry info for sector brief.

Clean up comm.c so that mantra and adrenaline rush are in their own functions instead of crammed into signal_char().

Fix Musashi's katana to use the proper gender in the spec.