Cleanup Cloud of Confusion code a little and remove the breakthrough check for Sphere; no other debuff spell similar to it (blindness, curse, warchant, etc.) has a breakthrough check, as they do not inflict damage. This spell also does not "attack" a target, as it does not initiate combat.
Modified Vampiric Touch to be usable by all alignments. The spell now shifts the caster's alignment 50 points towards evil per cast. This shift is greater (by up to 50 more) if the caster and the victim have positive alignments, and is based on how high the alignments are.

Removed Shroud of Existence.

Updates qf award to check if negative amounts are valid.
  Negative amounts that would result in the player having less than
  zero points are not awarded.