Update remort to close a "free exp" method by making remort always assume the character had an 18 con when leveling. This may be somewhat unfortunate for a "true" newbie, but so be it.
Change exp_table to be a const.

Change Combat Zen's chance to move while paralyzed to 50% chance.

Fix an issue with "score query exp_to_level", plus add a few more strings and showing help when a query was unknown.

General updates to "score query", as well as a new help entry for it.

A few fixes to other help entries.

Minor tweaks to donation room post-ID update.