Refactored quest_generate() some more. This cleans it up a bit and adds a couple sanity checks.
Added mob_disarm() to fight.c and fight.h. This returns a TRUE if the mob disarmed the PC, or FALSE otherwise. It has a bool arg that lets you specify whether to disarm to the ground; if not, the weapon is placed in the PC's inventory instead.

Updated a few specs that were clear copies of the mob attack disarm to use the new function.