Updated Banzai to remove the breakthrough check for invul on the physical hit portion, as that already checks for invul due to calling hit().
Updated Tranquility to now be a group spell. It costs 150 mana with a reduced duration of 6, but targets all members of the caster's group (not the room). It also has a special effect of re-applying the full duration on subsequent casts.

Fixed Talon spec in Vagabond cove so that it properly affects the group member when it is not the wielder of the talon.

Updated Ninja help files for formatting and clarity.