Further tweaks to AQ mobs based on feedback and tuning. Added and moved a few existing.
Made it so you can wake someone up that is affected by spell sleep if they are in a safe room. Emir and a few other mobs sleep you and then if you get recalled, you end up stuck in temple until the spell expires; if you weren't in a safe room, you could be pummeled awake (or some other attack/skill).

Changed group exp bonus sot hat it is based on multiples of 5 and only applies at 6+ instead of the multiples of 3 and starting at 4+ as pre-perma-BAM.

Modified a missed portion of [Blah's TARGET] further down in the show_char_to_char() code.