Updated aq mobs in Remorhaz, Golden Plaza, Marikith, and Moria 3
More revamp of generate_quest() (just style mostly).

Re-wrote most of generate_aq_order() in the same fashion as generate_quest(). No more "try x number of times and then give up", which should be a good thing. This function now builds a list of eligible vnums, shuffles the list, then picks the first 4 (the VNums will always be unique and valid, as long as the master AQ obj list does not contain errors or duplicates).

Added a couple new utility functions. I may add generic versions alter, if they are useful other places.

Missed the fact that auras go first and this needs to have strcat instead of strcpy.

Further tweaks to AQ mobs based on feedback and tuning. Added and moved a few existing.

Made it so you can wake someone up that is affected by spell sleep if they are in a safe room. Emir and a few other mobs sleep you and then if you get recalled, you end up stuck in temple until the spell expires; if you weren't in a safe room, you could be pummeled awake (or some other attack/skill).

Changed group exp bonus sot hat it is based on multiples of 5 and only applies at 6+ instead of the multiples of 3 and starting at 4+ as pre-perma-BAM.

Modified a missed portion of [Blah's TARGET] further down in the show_char_to_char() code.