Update Golden Pagoda - adds Uber Mustafa
Add Questy Boba for 2017 Christmas quest

Fixed help file entry for Blood Lust.

Removed the block from using parentheses in titles, as the likelihood of abusing "fake flags" is low. This is a policing problem if it ever happens.

Added a spec for Boba's santa on death to spawn a friendly version.