Fixed mithril hammer spec in Crystal Tower sot hat it only consumes one of each required item, not all of them. Cleaned up the code while I was at it.
Fixed Midgaard Mayor's specs up so that he now patrols town properly. He has some new emotes and commentary as well.

Added the ability to see the original room that an AQ mob was in when the quest was obtained, so you can recall this information with 'aquest info'; this is really just a QoL change in case players don't have an action set to store this info, or if they set temp variables and disconnect/crash/etc.