Update to new RNG.
Sometimes you decide you are a bit silly.

I re-added and updated the "old RNG" code to use as a fallback if RAND_bytes() ever fails. Of course, if it does, that should mean we're dead anyway...

The old RNG is restored for any "fallback" purposes. I also am keeping srand() in main() in case anyone decides to be silly and use rand(), assuming it's been seeded.

Remove the OLD_RNG build flags.