Added Night to IMP protections/extended functionality code.
Warlord and Gladiator have been completely re-worked. Warlord is the de-facto passive skill, AC-based tank with all the desirable tank skills. Gladiator is decidedly NOT a tank class, and has had all of its meaningful tank utility shifted away in favor of increased versatility, while also not adding significant damage over what Headbutt and Hostile already offer.

As part of the balancing of Warriors (as well as Commando), the base "magic numbers" for dual/triple have been modified. I have a spreadsheet that shows the old vs. the new rates, but the basic idea is that dual/triple/quad are boosted on the "low end", while their rates are "the same" on the top end. That is, a Gladiator today with 25 dex hits at the same rate as one will "tomorrow", but one with 18 dex will hit more often than it does now. This change is primarily to prevent Rush/Hostile from having such a huge swing (nearly 100% more hits) in the end than it does now.

The combat system has been further massaged and reworked, with some more stuff moved out of perform_hit(), as well as a simplified combat loop for dhit/thit/qhit. Further simplification needs to be done for skills that are "baked in" to perform_hit() at some point in the future.

The entirety of mobact.c has been re-written, while maintaining the "spirit and feel" of the original. This simplification resulted in a massive reduction in complexity and lines of code.

The mob_attack code in fight.c has been completely re-written and broken into smaller components. This is in the same vein as mobact.c and eventually the code should be moved to mobact.c out of fight.c for more clarity and code cleanliness.

Lots of code was touched after updating utils.h to "macroify" more of the direct struct references throughout the mud.

A few tweaks were made here and there to various other parts of the code, either as part of Warrior updates, or to help balance other aspects of the game after a review of changes made during the Warrior revamp.

Utility.c saw more work in the RNG portion, to allow for more options for calls to things like number() and dice(). An expanded func for each takes an option to allow generating normal numbers, averages, min and max numbers.

Fixed Lucifer so he will never fail to wield his sword (or, shouldn't, anyway).