Fix incorrect claw messaging for success_to_vict.
Actually, you know, check for SKILL_BLOCK, and stuff.

When combined with Hostile and 25 dex, compared to competing subclass skills (particularly Clobber), Headbutt is too strong as-is. Lowering the damage to keep Gladiator damage under control. This is still a significant improvement over Punch in the same time frame, even considering Punch has a higher success rate and frequency.