Fix a long standing bug in damage() for cold/sonic/blah damage types. Probably my fault, like 7 years ago.
Changed shadows so they mirror the weapon type of the Infidel. Some of the combat text code is pretty ugly, but maybe I'll fix it up someday along with the other fight.c revamps I've been working on.

Fixed an immortal-only bug with Dusk Requiem due to the way auto-collapsing shadows work.