Further AP tweaks.
Blood Lust mana cost was reduced to 70 (down from 75).

Rage mana cost was increased to 130 (up from 75), but also applies the Blood Lust spell automatically when successfully cast by Anti-Paladin characters. This is a quality of life change, primarily.

Shadowstep is slightly altered and moved to SC4. It is now slightly less effective in the dark for non-evil characters, but maintains potency for evil characters. Base damage was also reduced by 0.5x across the board, but is improved by SC5.

Bathed in Blood is moved to SC5. It now has more chances to affect the AP, through synergy with Spirit Levy, Desecrate, and Blood Lust procs. Additionally, Backstab can now potentially increase the blood level of the room at a 20% chance when used by an SC5 Defiler.

Fixed a bug where the chance of Vamp Touch on Backstab (Bathed in Blood perk) wasn't increasing due to blood level.

As noted above, Spirit Levy and Desecrate were altered to adjust the blood level of the room they are cast in by one, if the corpse targeted was from a creature no more than 10 levels below the caster. Blood Lust bites have a 20% chance of increasing the blood level of the room by one.

Fixed a bug where casting Shadow Wraith with not enough mana to pay the additional mana cost would result in a "can't summon enough energy" message, but the base mana cost (100) was still imposed. As a workaround, refund the base mana cost of the spell when this happens (which will result in 0 mana being consumed, as expected).