Updated Crusader subclass.
Blessing of Sacrifice now replaced Protect at Crusader SC2. This new spell heals all mortals in the caster's room for 333 hit points, at a cost of 100 mana and 100 hit points taken from the caster.

Focus is now a passive ability that grants the following benefits:
- Increased concentration rate by 10%
- Increases caster's alignment by 10 when casting heal, lay on hands, blessing of sacrifice, or power of faith
- Increases the chance of reduced mana cost for spells cast above level 41 by 10% (this adds to the alread-existing benefit for high level casters).

Commented out Charge stuff that was left behind in the code; I'll clean a lot of the commented sections up at a later date.