Adjusted Commando subclasses.
Iron Skin can now be cast on others, but only provides -10 AC, instead of the full -25 AC when self-cast. This helps non-tanking Mercenaries provide some useful benefit when in a group.  Mana cost and duration remain unchanged, making it fairly expensive for only -10 AC, but it's icing on the cake.

Devastation and Incendiary Cloud have swapped subclasses; their SC ranks remain unchanged.

Incendiary Cloud has been changed to inflict 700 damage initially, followed by 6 pulses of 100 damage, inflicted every ~ 10 seconds (MOB_ACT). Mana cost was reduced to 150, making this a more attractive spell to cast. Mana efficiency is good, but the payout is delayed and it would not be desirable to spam this spell.

Frost Bolt has been removed from Legionnaire and is now replaced with Snipe.

Snipe is a toggled skill and is similar to an automatic Disembowel, attempted every ~ 10 seconds (MOB_ACT). The chance of triggering is a fixed 20% chance. This skill inflicts 30x the commando's level (up to 1500 at level 50).