Only capitalize "Somebody" and revert to past behavior.
Added some hack code to utility.c to allow for mortals to see invis people in act() when specified with 'hide_invis' of -1. This is to allow for (particularly) the gf command to allow socials to show names of mortals, even if they are invis/impy/etc. Not the best way to do this, probably, but I didn't want to touch a ton of code at the moment (it'd be better to have an 'act_print' type function that can print to a variable, rather than directly to a descriptor, but that is for another day).

Fixed an oops where I did not unmerge half-finished cleric updates in spell_parser.c, thus removing clarity, degenerate, etc.

Fixed a cosmetic issue for immortals with invis/impy.