Added new antirent (30050)
Move Adept to 2 point AQ, as it was out of line with other 1 point mobs.

Fix the AFF2 for Rage in several places.

Remove perform_wear, as it's no longer moved.

Add a few comments to generate_quest, since I was looking through it and decided it needed some.

Fix a typo in spec.WOT.c.

Fix "0 Tick" by using != 1 instead of > 1 in do_affects

Re-write of do_wear and associated functions. This vastly cleans up the code and the "concept" behind it. It is also far more extensibile if we ever add new EQ slot types.

Added Fortification, Perceive, and Rage to show messages when gear is worn/removed, as well as in 'affect'.

Some misc fixes and formatting.