Fix the logic for two-slot items. Right now, only 2neck exists, but in the future maybe we'd have 2ring and/or 2wrist (or some other unknown slot). At present, because each eq item supports a max of 2 wear positions, we just need to search for the two-slot counterpart in the wear_info struct array and check if we're already wearing an item of that type in either slot. If we ever add a 3rd slot (like 3 fingers that can wear rings, etc.) then the whole shebang will need to obviously be overhauled to support that.
Anyway, the logic is a little funky here, but I tried to comment it so that it makes sense. Tested by both myself and Jearom in every combo we could think of, and this fixes the logic for neck and 2neck (the only two-slot item type we have right now).