More struggles to get Tremor/Rimefang better differentiated and more useful/unique. Overall, they are still pretty similar in several ways.
Updated Zeal to match help file for failure skill lag and removed the victim skill lag/mob attack delay because I felt it was too strong for a spammable skill with a very high success rate that hits everyone in the room. That said, this functionality was re-purposed into Rimefang, but it has a mana cost and saving throws.