Further tweaks to Rimefang, after doing some tests, plus a fix to a bug that I think has plagued it for years.
Further improvements/consistency for dynamic enchant names. Will need to come up with a better way (such as using #define in enchant.h, or similar) to store names so they don't have to be updated and verified all over the place, since enchant find is based on literal string matching.