Fixed oversight in practice code for Iron Skin.
Fixed an issue with do_kill() due to refactored logic.

Added a new attack type "TYPE_WEAPON2" that does what it says on the tin. This avoids the need to game even/odd hit numbers passed to perform_hit(), etc. It is a negative type number, so it can realistically only be used for "standard melee hits" (e.g. auto attacks).

Reversed 1st attempt at Leaf Spear "fix" that was specific to addressing Sidearm. I've been wanting to reduce Leaf Spear's proc effectiveness for a while, but wasn't sure what I needed to do. I've simply made a change so that it procs a single hit without the chance for dual/triple/quad. This levels the playing field for all classes that can wield it.