Stepped back and undid the still-funky logic in act_by_type(). There were still some issues. Testing some more, hopefully I got it this time. Sometimes you take the crazy pills.
Improved Trophy so that it doesn't decay corpse, as this messes with skinning. I wanted to free up an ITEM flag, since ITEM_NO_TROPHY is a bit of a waste. Setting the corpse to material type CLASS_BLOB after trophy works here, and is hilarious.

Reverted ZEN back to COMBAT ZEN for Mystic, since I never went through with that change, but it got left in the help file.