Fixed several minor bugs in the new comm system.
Fixed some cosmetic issues with subclass master changes.

Restored previous functionality for generic_find(), which seems a bit odd to me, but "broke" some input unexpectedly. The find aborts if there are args beyond what is being searched for. The most noticeable was when trying to do "open stone up" if you had a lucky stone. Previously, it would abort the search, because of the "up" text after "stone" and would instead look for exits to open, but it would target the lucky stone if you only typed "open stone", so I restored this behavior.

Several in-progress improvements in handler and interpreter.c code.

Fixed fill word filtering in one_argument and associated functions, that I apparently broke long ago. Luckily it's pretty foreign to try to "open the stone in up" for most players.