Fix unfinished code that I didn't intend to commit, but did, because I was running late for dinner.
Various minor updates for some low-level changes planned in the near future.

Fixed several minor bugs in the new comm system.

Fixed some cosmetic issues with subclass master changes.

Restored previous functionality for generic_find(), which seems a bit odd to me, but "broke" some input unexpectedly. The find aborts if there are args beyond what is being searched for. The most noticeable was when trying to do "open stone up" if you had a lucky stone. Previously, it would abort the search, because of the "up" text after "stone" and would instead look for exits to open, but it would target the lucky stone if you only typed "open stone", so I restored this behavior.

Several in-progress improvements in handler and interpreter.c code.

Fixed fill word filtering in one_argument and associated functions, that I apparently broke long ago. Luckily it's pretty foreign to try to "open the stone in up" for most players.

Improved drunkify_string so that now it is a generic function that can be used to drunkify any text. Parameter order was changed so that it follows the same format as snprintf, and the like, with the required 'drunk_level' variable. Accepts ... args as well. This should probably get moved somewhere else, but I need to find a suitable place.

Spruced up the comm code a little more.

Fixed do_kill() to the old behavior, so IMP raw kill is after weapon action descriptions; gotta keep Sane and his old hickory stick happy after all.