Another fix to 2H weapon wield logic. Planning to re-write wear/remove logic again in the future to make it more robust, and also to allow targeting specific slots with the wear/wield commands.
Remove order items that are either broken (e.g. don't actually load), or that are junked 99.99% of the time. Add some different order items in to compensate and for flavor.

Adjust snipe enchantment to use GET_DISP_NAME() instead of GET_NAME() in case a mob uses it.

Adjust Unholy Blade of Sacrifice spec to make it worth using. 27 avg damage weapons aren't what they used to be.