Re-worked the zone/room/mob/obj allocation functions. Added binary search functions for each table.
Added allocation function for shops, which are now stored in a sorted list and can be searched via its own binary search function. Previously, shops were an unordered list. I didn't search for an update every place that searching shops is performed.

Token distribution function re-written to generate and cache a table of token mobs. The old method of a while loop that gave up after 1000 tries annoyed me, even though this is more complex.

Rashgugh teleportation function re-written to generate and cache a table of teleport rooms. The old method could and would cause the game to hang in a while loop if no valid rooms were loaded in the game.

Added seasons and seasonal day/night cycles. Some groundwork (commented out) for biomes and zone-specific weather. On the list of things to add are temperatures and moon phases. Low priority work though.

Tweaked some of the utility functions in utility.c for consistency and correctness.