Attempted to reconcile reported light issues.
Ensure that what a light source is held by a player, the light counter in the room is incremented.

Ensure that when a light source burns out, the light counter in the room is decremented.

Allow mobs to see in the dark without infravision (they already can attack/spec in the dark, so it doesn't really matter, except for when immortals force them to take actions).

Add the ability to see the light counter in rooms using 'stat room', and also show if the room is otherwise lit (by the sun, room flags, etc). Useful in general, and also for debugging.