Adjust remort multiplier for non-hybrid classes.
Lowercase something that was uppercase for no reason in spec.ecanyon.c.

Fix sting() spec in spec.vanity.c so you can't spam backstab during fight to cause it to proc. This was fixed in spec.marikith.c, but never made its way to custom gear.

Adjust Power of Devotion so that the sanctuary affect is always put in place. This allows sanc effect to persist if the player already had sanctuary via the Sanctuary spell, and it wore off.

Adjust Power of Faith to min heal 375 for 50 mana and max heal for 1500 for 125 mana.