Re-wrote the vast majority of the affect and enchant system. Moved associated functionality into aff_ench.c it's header file. Nearly all of the legacy enchant.c code is now a wrapper for the new code. The same is true of the affect-related code in handler.c.
The affect/enchant system now has the ability to place affects/enchants in their respective lists based on a calculated priority. As of now, this is a fixed ordering via list, but could be modified in the future.

Created a new Enchantment structure (struct enchantment_t), which is extended in a few ways. Temporary storage is expanded to an array of 4 integers, as well as a new character pointer for arbitrary metadata. As usual, data other than what the AFF can contain is not stored on the pfile, and is transitory. The ENCH typedef now an alias for the new structure.

Added a new function pointer to struct char_data. For players, this function is hooked at login and points to a special character function in the new char_spec.c file and its associated header. For mobiles, this can be dynamically assigned, such as via spec. This allows for players to receive all signals sent via the comm.c signal passing code, and lets us hook any kind of signal-processing functions without having to insert code into comm.c, fight.c, or other places where we were doing so previously. Automatic skills such as snipe, shadowstep, or adrenaline are hooked here. For mobiles, we can now dynamically assign specs on a per-mobile basis, which is less useful, but still an option.

Converted many things that were hard-coded in comm.c or fight.c to use the new special character function as described above, as well as converted some of these to enchantments (such as Mantra and Wither).

Mantra heals the same amount of hit points over a slightly shorter period (90 seconds instead of 100). Wither functionality was tuned, refined, and made to correctly reflect the help file entry.

Tweaked some of the combat code, removing some undesired behavior, such as NPCs receiving hitroll/damroll bonuses from strength (all mobs have fixed stats, and strength is fixed at 18/0, which was providing a small bonus).

Made sprinttype and sprintbit functions safe and renamed them to snprintf_type and snprintf_bits respectively. Re-wired existing calls to the new functions. The old calls are commented out for now, and will be removed in a future update when we've made sure nothing was missed or lost in translation.

Added a new message, MSG_VIOLENCE_POST_HIT, which is sent after the melee attack in perform_violence (MSG_VIOLENCE is sent before it). This is to allow for more precise timing of certain specs as part of combat.

Made a few tweaks to the Demon Invasion specs.

Fixed or tweaked several minor things for correctness or tuning.

There's still some more functionality tweaks for enchantments to do in the future, as well as removing ranks as enchants entirely, but that will come in a later update.