Update darkened shadow orb (14053)   + now useable by all   + introduce half pulse lag both in and out of combat   + reduce hp drain by half
Update shadow orb
  + introduce half pulse lag both in and out of combat

Clean up do_look a little more.

Added "look rdesc" as a way to explicitly print the room description, if for example a player wants to see it without toggling brief on/off.

Added "display ticker" as a way to print the message "TICK!" to the player on tick, instead of having to rely on objects or other methods.

Fix PvP check.