Update remort system to allow the purchase of alignment tokens (good/evil). If no alignment token is purchased, the character's alignment will be set to 0 (neutral).
Repurposed Avatar and Name tokens for the new alignment tokens, as they were disabled anyway.

Re-enabled the ability to remort into a Nomad if a player so chooses. The Nomad class is still disabled for new players at this time.

Added the ability for players to disable/rename their remort enchantment, thus halting/resuming earning of remort experience.