Add files via upload
Changed the stats of gobbles to last longer.   Also updated the pilgrim leader since he is very important to the zone being available.   Modified a lot of objects to decay to keep the zone cleaner and to ensure that you cant hoard items.   

Upped the XP values of the turkeys to be competitive with end game mobs as well.  

Signed-off-by: Aaron Grimmer <87430970+Arodtanjoe@users.noreply.github.com>