Add files via upload
Removed Christmas Only Checks so they can be enjoyed for the month of January- Will replace once the event is over.  This way it is active next year only on Christmas.  

Cleaned up the blank space and commented out code. 

Also updated specs for items to be more in line with current loading gear. 

Signed-off-by: Aaron Grimmer <87430970+Arodtanjoe@users.noreply.github.com>